Expanding the tutorial notebook for SF. Moved the `SF_METHODS` variable into a new module, added tests for that.
